introduce-metrics-reporter-move: Move metrics reporting to SchedulerMetricsReporter and retire metrics mixin
Mechanical cut + paste for the ``introduce-metrics-reporter`` mech move.

Cut the methods (all @staticmethod after prep) +
``PrefillStats`` dataclass + module constants
(``RECORD_STEP_TIME`` / ``LOG_FORWARD_ITERS`` /
``ENABLE_METRICS_DEVICE_TIMER``) from
``SchedulerMetricsMixin`` and paste them into the
``SchedulerMetricsReporter`` class body / module scope at
``scheduler_components/metrics_reporter.py``. Drop ``@staticmethod``
decorators; simplify ``self: "SchedulerMetricsReporter"`` annotation
to bare ``self`` (in class context the type is implicit). Strip the
``SchedulerMetricsMixin.<sibling>(self, ...)`` qualified form on
internal sibling calls (and the matching ctor-body
``SchedulerMetricsMixin.init_metrics(self, ...)`` /
``SchedulerMetricsMixin.install_device_timer_on_runners(self)``) — the
methods are now in the same class, plain ``self.<method>(...)`` resolves.
Method bodies otherwise byte-identical to the post-prep state.

Source mixin module retired: the file
``observability/scheduler_metrics_mixin.py`` is removed.
``SchedulerMetricsMixin`` is dropped from ``Scheduler`` parents.

All callers updated:
  ``self.<method>(self.metrics_reporter, ...)`` →
  ``self.metrics_reporter.<method>(...)``
(pure prefix transformation, via the canonical
``rewrite_method_call_site`` helper):
- ``scheduler.py``: hot-path sister calls + the
  ``self.metrics_reporter.log_batch_result_stats(...)`` etc. variants.
- ``scheduler_output_processor_mixin.py``:
  ``report_prefill_stats`` / ``update_spec_metrics`` /
  ``report_decode_stats``.
- ``disaggregation/prefill.py``: ``report_prefill_stats``.
- ``dllm/mixin/scheduler.py``: ``report_prefill_stats`` + PrefillStats
  import path rewrite.

Import-path rewrites (pure prefix replace):
- ``scheduler.py``: ``RECORD_STEP_TIME`` / ``PrefillStats`` import block
  moves from observability/scheduler_metrics_mixin → scheduler_components/metrics_reporter.
- ``schedule_batch.py``: TYPE_CHECKING ``PrefillStats`` import.
- ``dllm/mixin/scheduler.py``: ``PrefillStats`` local import.

Refactor chain ID: introduce-metrics-reporter-move
